feat Multi-target skills
- Handle Phase transition by Event
  - combat::mod
    - add event
  - combat::phases
    - phase_transition():
    event handler
  - use event to phase transi
    - ui::character_sheet
    - ui::combat_system
    - ui::player_interaction
- Possibility to target multiple depending to the skill selected
  - combat::alterations & combat::skills
    - alteration/skill def
      - tager_number: usize
  - combat::mod
    - Action def
  - combat::phases
    - execution_phase()
      - exec the skill to all the target
  - ui::combat_system
    - update_targeted_unit()
      - allow the caster to be reselected after completing an action with him or her
      - If the number of targets is > of max possible:
      Remove the first and push the new one
  - ui::player_interaction
    - confirm_action_button()
      - Upon pressing of the new button: `ConfirmActionButton`
        - If the number of currently targeted entity.ies is = the number required:
        phase transi/confirm the action
        - Else: 'print' "NOT ENOUGH TARGETS"
    - docs
- Multi-target skills Example
  - combat::skill_list
    - implosion()
    `Deal 25dmg to 3 targets`
  - ui:combat_panel
    - create the implosion skill button
- combat::skill::execute_skill()
  - Update docs
  - Update Logs
- deprecate tag
  - ui::combat_system
    - target_random_system()

+++ src/combat/alterations.rs
@@ -55,7 +55,7 @@ pub struct Alteration {
     ///
     /// - target all ally party: (Ally, 6)
     /// - self-target: (Ally, 0)
-    pub target_option: (TargetSide, i32),
+    pub target_option: (TargetSide, usize),
 
     /// hp dealt or healed each time the target plays
     ///
